# swagger.json
{"swagger":"2.0","info":{"description":"API for Accessing Netflix Pricing","title":"Netflix Pricing Service API","termsOfService":"Contact Pricing Data Access API team if you are using the API"},"host":"localhost:8080","basePath":"/","tags":[{"name":"Customers API","description":"Not the main api - used for dummy customer population and verification"},{"name":"Netflix Pricing Customers API","description":"Main REST API for accessing & rollout out Price"}],"paths":{"/customers":{"get":{"tags":["Customers API"],"summary":"getAllCustomersOrByCountryCode","operationId":"getAllCustomersOrByCountryCodeUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"countryCode","in":"query","description":"countryCode","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Iterable«Customer»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["Customers API"],"summary":"createDummyCustomersForS1PS","operationId":"createDummyCustomersForS1PSUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/customers/{COUNTRY_CODE}/{SERVICE_PLAN}/{PRICE}/{CURRENCY}/{COUNT}":{"post":{"tags":["Customers API"],"summary":"createDummyCustomersForCountryServicePlan","operationId":"createDummyCustomersForCountryServicePlanUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"COUNTRY_CODE","in":"path","description":"COUNTRY_CODE","required":true,"type":"string"},{"name":"SERVICE_PLAN","in":"path","description":"SERVICE_PLAN","required":true,"type":"string"},{"name":"PRICE","in":"path","description":"PRICE","required":true,"type":"number"},{"name":"CURRENCY","in":"path","description":"CURRENCY","required":true,"type":"string"},{"name":"COUNT","in":"path","description":"COUNT","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/customers/{ID}":{"get":{"tags":["Customers API"],"summary":"getCustomerById","operationId":"getCustomerByIdUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"ID","in":"path","description":"ID","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Optional«Customer»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/prices":{"get":{"tags":["Netflix Pricing Customers API"],"summary":"getAllServicePlanPricesForAllCountries","operationId":"getAllServicePlanPricesForAllCountriesUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"active","in":"query","description":"active","required":false,"type":"boolean","default":false}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ServicePlanPriceDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["Netflix Pricing Customers API"],"summary":"updatePrices","operationId":"updatePricesUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"servicePlanPriceDTO","description":"servicePlanPriceDTO","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/ServicePlanPriceDTO"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Iterable«ServicePlanPriceDTO»"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["Netflix Pricing Customers API"],"summary":"deletePriceChangeEntryIfNotActive","operationId":"deletePriceChangeEntryIfNotActiveUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"query","description":"id","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/prices/{COUNTRY}":{"get":{"tags":["Netflix Pricing Customers API"],"summary":"getServicePlanPriceByCountry","operationId":"getServicePlanPriceByCountryUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"COUNTRY","in":"path","description":"COUNTRY","required":true,"type":"string"},{"name":"active","in":"query","description":"active","required":false,"type":"boolean","default":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ServicePlanPriceDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/prices/{COUNTRY}/{PLAN}":{"get":{"tags":["Netflix Pricing Customers API"],"summary":"getServicePlanPriceByCountryAndServicePlan","operationId":"getServicePlanPriceByCountryAndServicePlanUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"COUNTRY","in":"path","description":"COUNTRY","required":true,"type":"string"},{"name":"PLAN","in":"path","description":"PLAN","required":true,"type":"string","enum":["SP1S","SP2S","SP4S"]},{"name":"active","in":"query","description":"active","required":false,"type":"boolean","default":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ServicePlanPriceDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Customer":{"type":"object","properties":{"country":{"type":"string"},"currency":{"type":"string"},"id":{"type":"string"},"price":{"type":"number"},"servicePlan":{"type":"string"}}},"Iterable«Customer»":{"type":"object"},"Iterable«ServicePlanPriceDTO»":{"type":"object"},"Optional«Customer»":{"type":"object","properties":{"present":{"type":"boolean"}}},"ServicePlanPriceDTO":{"type":"object","properties":{"active":{"type":"boolean"},"countryISOCode":{"type":"string"},"currency":{"type":"string"},"effectiveDate":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"price":{"type":"number"},"servicePlan":{"type":"string","enum":["SP1S","SP2S","SP4S"]}}}}}